+++ include/leds/strip.h
