+++ src/leds/power.cpp
@@ -1,4 +1,4 @@
-#include "leds/handler.h"
+#include "leds/power.h"
 #include "subsystems.h"
 
 void LedPowerHandler::update() {
